in progress
new home page for v2, moved glob to separate component and added to backdrop
