+++ biome.json
@@ -10,6 +10,7 @@
 		"useIgnoreFile": true
 	},
 	"files": {
+		"ignoreUnknown": true,
 		"ignore": ["./components/ui/**"]
 	},
 	"linter": {
