+++ package.json
@@ -29,6 +29,8 @@
 		"@radix-ui/react-toast": "^1.2.4",
 		"@radix-ui/react-toggle": "^1.1.1",
 		"@radix-ui/react-tooltip": "^1.1.6",
+		"@react-three/drei": "^9.120.4",
+		"@react-three/fiber": "^8.17.10",
 		"@vercel/analytics": "^1.4.1",
 		"@vercel/speed-insights": "^1.1.0",
 		"autoprefixer": "10.4.20",
@@ -49,17 +51,20 @@
 		"tailwind-merge": "^2.6.0",
 		"tailwindcss": "^3.4.17",
 		"tailwindcss-animate": "^1.0.7",
+		"three": "^0.171.0",
+		"three-stdlib": "^2.35.2",
 		"typescript": "^5.7.2",
 		"typewriter-effect": "^2.21.0",
 		"zod": "^3.24.1"
 	},
 	"devDependencies": {
-		"@types/react-google-recaptcha": "^2.1.9",
+		"@biomejs/biome": "^1.9.4",
 		"@types/node": "^22.10.2",
 		"@types/nodemailer": "^6.4.17",
 		"@types/react": "19.0.2",
 		"@types/react-dom": "19.0.2",
-		"@biomejs/biome": "^1.9.4",
+		"@types/react-google-recaptcha": "^2.1.9",
+		"@types/three": "^0.171.0",
 		"eslint": "8.57.0",
 		"eslint-config-next": "15.1.2"
 	},
